Refactoring 1.5 Game logic replaced to Server.

+++ BlackJack.Entities/Step.cs
@@ -10,6 +10,6 @@ namespace BlackJack.Entities
     {
         public long WinnerId { get; set; }
         public long GameId { get; set; }
-        public long GameProcess { get; set; }
+        public int GameProcess { get; set; }
     }
 }
